net_quest/requirements.txt
jinja2 
oauth2client  
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib 
httplib2 
python-docx 
XlsxWriter 
